% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split-testing.R
\name{conversion_expected_loss}
\alias{conversion_expected_loss}
\title{Find the expected loss in a conversion experiment}
\usage{
conversion_expected_loss(
  sA,
  nA,
  sB,
  nB,
  prior_alpha = 1,
  prior_beta = 1,
  approx = FALSE,
  ...
)
}
\arguments{
\item{prior_alpha}{prior parameter to use for alpha (default uniform)}

\item{prior_beta}{prior parameter to use for beta (default uniform)}

\item{approx}{whether to use a normal approximation for h}

\item{...}{Extra arguments, passed on to \code{\link{expected_loss}}}

\item{success_A}{number of successes (e.g. clicks) in condition A}

\item{total_A}{total number of exposures in A}

\item{success_B}{number of successes (e.g. clicks) in condition B}

\item{total_B}{total number of exposures in B}
}
\description{
This is a wrapper around \code{\link{expected_loss}}. It applies
a prior and calculates in terms of successes and totals of A and B,
instead of the beta distribution. Note that the "error" is treatment
A being better than B.
}
