% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math-functions.R
\name{expected_loss}
\alias{expected_loss}
\title{Calculate the expected loss from assuming the wrong beta}
\usage{
expected_loss(a, b, c, d, approx = FALSE, ...)
}
\arguments{
\item{a}{alpha parameter for first Beta}

\item{b}{beta parameter for second Beta}

\item{c}{alpha parameter for first Beta}

\item{d}{beta parameter for second Beta}

\item{approx}{whether to use a normal approximation when calculating h}

\item{...}{Extra arguments, ignored}
}
\description{
Find the expected loss from two Beta posteriors if one assumes (wrongly)
that a / b > c / d. That is, \code{E[max(Beta(c, d) - Beta(a, b), 0)]}. This comes from
https://www.chrisstucchio.com/blog/2014/bayesian_ab_decision_rule.html. Note
that the \code{approx = TRUE} version is vectorized, but the exact
version is not.
}
