% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math-functions.R
\name{h}
\alias{h}
\title{Calculate the probability one random beta variable is greater than another}
\usage{
h(a, b, c, d, approx = FALSE, log_h = FALSE)
}
\arguments{
\item{a}{alpha parameter for first Beta}

\item{b}{beta parameter for second Beta}

\item{c}{alpha parameter for first Beta}

\item{d}{beta parameter for second Beta}

\item{approx}{whether to use a normal approximation to the beta}

\item{log_h}{whether to return \code{log(h(a, b, c, d))} rather than
\code{h(a, b, c, d)}}
}
\description{
Find the probability that \code{Beta(a, b) > Beta(c, d)}. This comes from
https://www.chrisstucchio.com/blog/2014/bayesian_ab_decision_rule.html. Note
that the \code{approx = TRUE} version is vectorized, but the exact version
is not.
}
