% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{perform_simulation}
\alias{perform_simulation}
\title{Perform a simulation of A/B testing based on a data frame of parameters}
\usage{
perform_simulation(params, days = 20, per_day = 100, proportion_A = 0.1, ...)
}
\arguments{
\item{params}{A data frame of parameters, each of which will be passed on to
\code{\link{conversion_expected_loss}}}

\item{days}{Number of days to run}

\item{per_day}{Number of "clicks" per day; overridden if present in params}

\item{proportion_A}{"Clickthrough rate" in treatment A}

\item{...}{Extra arguments, passed on to \code{\link{conversion_expected_loss}}}
}
\description{
Perform a simulation of A/B testing based on a data frame of parameters
}
\examples{

library(dplyr)

sim <- tibble(replicate = seq_len(nreps)) \%>\%
  mutate(proportion_A = .001, effect = 0, per_day = 10000) \%>\%
  perform_simulation()

library(ggplot2)
ggplot(sim, aes(day, expected_loss, group = replicate)) +
  geom_line(alpha = .5) +
  scale_y_log10()

}
