% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vectorized-prop-test.R
\name{vectorized_prop_test_approx}
\alias{vectorized_prop_test_approx}
\title{A vectorized version of the prop.test function}
\usage{
vectorized_prop_test_approx(a, b, c, d)
}
\description{
A vectorized version of the prop.test function
}

