% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vectorized-prop-test.R
\name{vectorized_prop_test_exact}
\alias{vectorized_prop_test_exact}
\title{Perform a Fisher exact test for each set of a, b, c, d}
\usage{
vectorized_prop_test_exact(a, b, c, d)
}
\description{
Perform a Fisher exact test for each set of a, b, c, d
}

